Log and maxAllowedFood considering treaties+messages
