+++ config.json
@@ -8,7 +8,7 @@
     "Team3Agents": 0,
     "Team4Agents": 0,
     "Team5Agents": 0,
-    "Team6Agents": 2,
+    "Team6Agents": 3,
     "Team7Agents": 0,
     "RandomAgents": 0,
     "AgentHP": 100,
